+++ lib/savon/core_ext/array.rb
@@ -16,8 +16,9 @@ module Savon
         each_with_index do |item, index|
           attrs = tag_attributes attributes, index
           case item
-            when ::Hash then xml.tag!(key, attrs) { xml << item.to_soap_xml }
-            else             xml.tag!(key, attrs) { xml << (escape_xml ? item.to_soap_value : item.to_soap_value!) }
+            when ::Hash   then xml.tag!(key, attrs) { xml << item.to_soap_xml }
+            when NilClass then xml.tag!(key, "xsi:nil" => "true")
+            else               xml.tag!(key, attrs) { xml << (escape_xml ? item.to_soap_value : item.to_soap_value!) }
           end
         end